feat: handle translation on the frontend
