+++ types/transcription.ts
@@ -41,3 +41,5 @@ export interface AudioChunk {
   data: ArrayBuffer;
   timestamp: number;
 }
+
+
